// apps/web/src/lib/api/generated/types/customMovieResponseDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Movies App API
 * REST API for Movies application

**Links:**
- [OpenAPI JSON](/api/docs/json)
- [OpenAPI YAML](/api/docs/yaml)
 * OpenAPI spec version: 1.0.0
 */
import type { CustomMovieResponseDtoStatus } from './customMovieResponseDtoStatus';

export interface CustomMovieResponseDto {
	id: number;
	groupId: number;
	title: string;
	posterPath?: string;
	overview?: string;
	releaseYear?: number;
	runtime?: number;
	/** Movie status */
	status: CustomMovieResponseDtoStatus;
	plannedDate?: string;
	watchedDate?: string;
	createdById: number;
	createdAt: string;
	updatedAt: string;
}